[kernel] Build everything with -Wconversion [part 25]
Changes to be committed:
	modified:   include/tilck/kernel/fs/vfs.h
	modified:   kernel/fs/fs_syscalls.c
	modified:   kernel/fs/mp.c

+++ kernel/fs/mp.c
@@ -109,7 +109,7 @@ mountpoint *mountpoint_get_next(_mp_cursor *c)
    ASSERT(c->curr_mp >= 0);
    ASSERT(!is_preemption_enabled());
 
-   for (u32 i = c->curr_mp++; i < ARRAY_SIZE(mps); i++) {
+   for (int i = c->curr_mp++; i < (int)ARRAY_SIZE(mps); i++) {
       if (mps[i] != NULL)
          return mps[i];
    }
